+++ docs/release-notes/0.16.0.md
@@ -4,7 +4,7 @@
 ```
 * Reworked GPU {func}`~rapids_singlecell.tl.rank_genes_groups` Wilcoxon onto dedicated nanobind CUDA kernels {pr}`636` {smaller}`S Dicks`
 * {func}`~rapids_singlecell.tl.rank_genes_groups` no longer truncates gene names longer than 50 characters in ``uns[...]['names']`` (the field is now ``object`` dtype, matching Scanpy) {pr}`636` {smaller}`S Dicks`
-* Add {class}`~rapids_singlecell.ptg.Mixscape` for GPU-accelerated Mixscape (`perturbation_signature`, `mixscape`, `mixscale`, `lda`) {pr}`688` {smaller}`S Dicks`
+* Add {class}`~rapids_singlecell.ptg.Mixscape` (`perturbation_signature`, `mixscape`, `lda`) and {class}`~rapids_singlecell.ptg.Mixscale` (`perturbation_signature`, `mixscale`) for GPU-accelerated perturbation-efficacy analysis {pr}`688` {smaller}`S Dicks`
 
 ```{rubric} Performance
 ```
