+++ src/rapids_singlecell/pertpy_gpu/__init__.py
@@ -2,4 +2,5 @@ from __future__ import annotations
 
 from ._distance import Distance, MeanVar
 from ._guide_assignment import GuideAssignment
+from ._mixscale import Mixscale
 from ._mixscape import Mixscape
